+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Per connection http provider configuration
 
 ### Changed
+- `K8s.Conn.from_file/2` now returns an ok or error tuple
+- `K8s.Conn.from_service_account/N` now returns an ok or error tuple
 - `K8s.Conn.t()` is now the first argument in all runners. `K8s.Operation.t()` is now the second.
 - deprecated K8s.http_provider/0
 - deprecated K8s.Discovery.default_opts/0